requirements.txt
streamlit
pandas
numpy
scikit-learn
plotly
openpyxl
XlsxWriter
Pillow